+++ src/cli.ts
@@ -1,5 +1,6 @@
 #!/usr/bin/env node
 import assert from "node:assert/strict";
+import { readFileSync } from "node:fs";
 
 import {
   ArgumentParser,
@@ -8,12 +9,19 @@ import {
 } from "argparse";
 
 import { launch } from "./browser.js";
-import { defaultArgs, runConfigForArgs } from "./config.js";
+import { defaultLaunchArgs, runConfigForArgs } from "./config.js";
 import { getLogger, LoggingLevel } from "./logging.js";
 import { measureURL } from "./measure.js";
 import { BrowserType, MeasurementType } from "./types.js";
 
+const packageText: string = readFileSync("./package.json", "utf8");
+assert(typeof packageText === "string");
+const packageData = JSON.parse(packageText) as { version: string };
+const packageVersion: string = packageData.version;
+assert(packageVersion);
+
 const isDebugMode = process.env.PERF_TESTS_DEBUG === "1";
+const defaultArgs = defaultLaunchArgs();
 
 const parser = new ArgumentParser({
   description: "Run performance tests for a playwright version of a browser.",
@@ -21,7 +29,7 @@ const parser = new ArgumentParser({
 });
 parser.add_argument("-b", "--browser", {
   choices: Object.values(BrowserType),
-  default: BrowserType.Brave,
+  default: defaultArgs.browser,
   help: "Which browser family to use for this test.",
 });
 parser.add_argument("-d", "--user-data-dir", {
@@ -37,14 +45,14 @@ parser.add_argument("-d", "--user-data-dir", {
 });
 parser.add_argument("-l", "--logging", {
   choices: Object.values(LoggingLevel),
-  default: LoggingLevel.Info,
+  default: defaultArgs.loggingLevel,
   help:
     "What level of information to include when printing information during " +
     "the measurement.",
 });
 parser.add_argument("-m", "--measurements", {
   choices: Object.values(MeasurementType),
-  default: Object.values(MeasurementType),
+  default: defaultArgs.measurements,
   help:
     "Which measurements of performance to collect. By default, performs all " +
     "measurements.",
@@ -66,6 +74,15 @@ parser.add_argument("-p", "--profile", {
     "For Chromium, this is the name of the profile in an existing " +
     "--user-data-dir directory to load. (Only used for Chromium browsers)",
 });
+parser.add_argument("-r", "--preserve-pages", {
+  action: "store_true",
+  default: defaultArgs.preservePages,
+  help:
+    "Preserve and reopen any pages and tabs that are set as open in the " +
+    "specified --user-data-dir (if any). By default the tool will prevent " +
+    "any pages from being automatically opened, but if this argument is set, " +
+    "then any existing pages will be reopened before taking any measurements.",
+});
 parser.add_argument("-s", "--seconds", {
   default: defaultArgs.seconds,
   help: "Number of seconds to wait while measuring page performance.",
@@ -86,6 +103,10 @@ parser.add_argument("-u", "--url", {
   required: true,
   type: URL,
 });
+parser.add_argument("-v", "--version", {
+  action: "version",
+  version: packageVersion,
+});
 parser.add_argument("-x", "--binary-path", {
   help:
     "Path to the browser binary to run the measurements with. Note that " +
@@ -94,12 +115,12 @@ parser.add_argument("-x", "--binary-path", {
     "while the gecko and webkit ones do.",
 });
 parser.add_argument("--height", {
-  default: defaultArgs.viewport.height,
+  default: defaultArgs.viewport?.height,
   help: "The height of the browser viewport to use when loading pages.",
   type: "int",
 });
 parser.add_argument("--width", {
-  default: defaultArgs.viewport.width,
+  default: defaultArgs.viewport?.width,
   help: "The width of the browser viewport to use when loading pages.",
   type: "int",
 });
@@ -109,7 +130,8 @@ try {
   assert(rawArgs instanceof Namespace);
 
   const runConfig = await runConfigForArgs(rawArgs);
-  const { measurements, url, seconds, timeout, loggingLevel } = runConfig;
+  const { measurements, url, seconds, timeout } = runConfig;
+  const { loggingLevel, preservePages } = runConfig;
   const logger = getLogger(loggingLevel);
   const browserContext = await launch(logger, runConfig);
   const results = await measureURL(
@@ -119,6 +141,7 @@ try {
     seconds,
     timeout,
     measurements,
+    preservePages,
   );
 
   runConfig.output.write(JSON.stringify(results));