+++ src/config.ts
@@ -20,14 +20,20 @@ const { R_OK, W_OK, X_OK } = constants;
 const programName = "privacy-perf-comparisons";
 const validSchemes = ["http:", "https:"];
 
-export const defaultArgs = {
-  profile: "Default",
-  seconds: 30,
-  timeout: 30,
-  viewport: {
-    height: 1024,
-    width: 1280,
-  },
+export const defaultLaunchArgs = (): Partial<RunConfig> => {
+  return {
+    browser: BrowserType.Chromium,
+    loggingLevel: LoggingLevel.Info,
+    measurements: [MeasurementType.Network, MeasurementType.Timing],
+    preservePages: false,
+    profile: "Default",
+    seconds: 30,
+    timeout: 30,
+    viewport: {
+      height: 1024,
+      width: 1280,
+    },
+  };
 };
 
 const _fileCheck = async (
@@ -192,16 +198,8 @@ export const runConfigForArgs = async (args: Namespace): Promise<RunConfig> => {
   const isChromium =
     args.browser === BrowserType.Chromium || args.browser === BrowserType.Brave;
 
-  let userDataDir: undefined | string;
-  if (typeof args.user_data_dir === "string") {
-    userDataDir = args.user_data_dir;
-  }
-
   assert(typeof args.profile === "string");
-  if (
-    args.browser === BrowserType.Gecko &&
-    args.profile !== defaultArgs.profile
-  ) {
+  if (args.browser === BrowserType.Gecko && args.profile !== args.profile) {
     throw new Error(
       "Cannot use different profiles within the same profile " +
         "directory for Gecko browsers. Use --user-data-dir to use different " +
@@ -230,17 +228,23 @@ export const runConfigForArgs = async (args: Namespace): Promise<RunConfig> => {
 
   // This is not the most concise way to check these cases, but meant
   // to be the easiest to follow, by matching the above exact criteria
-  let isUserDataDirExisting: boolean | undefined;
-  let isProfileReadable: boolean | undefined;
-  if (!userDataDir) {
-    isUserDataDirExisting = false;
-    isProfileReadable = false;
-  } else if (await isPathToReadableDir(userDataDir)) {
-    isUserDataDirExisting = true;
-    isProfileReadable = await isPathToReadableDir(userDataDir, args.profile);
+  let userDataDirArg: undefined | string;
+  if (typeof args.user_data_dir === "string") {
+    userDataDirArg = args.user_data_dir;
   }
 
-  const isCaseOne = !userDataDir;
+  let isUserDataDirExisting: boolean | undefined;
+  let isProfileReadable: boolean | undefined;
+
+  if (!userDataDirArg) {
+    isUserDataDirExisting = false;
+    isProfileReadable = false;
+  } else if (await isPathToReadableDir(userDataDirArg)) {
+    isUserDataDirExisting = true;
+    isProfileReadable = await isPathToReadableDir(userDataDirArg, args.profile);
+  }
+
+  const isCaseOne = !userDataDirArg;
   const isCaseTwo = !isCaseOne && !isUserDataDirExisting;
   const isCaseThree =
     !isCaseOne && !isCaseTwo && !isChromium && isUserDataDirExisting;
@@ -260,20 +264,20 @@ export const runConfigForArgs = async (args: Namespace): Promise<RunConfig> => {
     validatedUserDataDir = tempDirPath.path;
     logMsg("\t", "- creating temporary user-data dir: ", validatedUserDataDir);
   } else if (isCaseTwo) {
-    validatedUserDataDir = userDataDir;
+    validatedUserDataDir = userDataDirArg;
     logMsg("\t", "- creating new user-data dir: ", validatedUserDataDir);
   } else if (isCaseThree) {
-    validatedUserDataDir = userDataDir;
+    validatedUserDataDir = userDataDirArg;
     logMsg("\t", "- using existing user-data dir: ", validatedUserDataDir);
   } else if (isCaseFour) {
-    assert(userDataDir);
-    validatedUserDataDir = join(userDataDir, args.profile);
+    assert(userDataDirArg);
+    validatedUserDataDir = join(userDataDirArg, args.profile);
     validatedProfile = args.profile;
     logMsg("\t", "- using user-data dir: ", validatedUserDataDir);
     logMsg("\t", "- with profile name: ", args.profile);
   } else {
     throw new Error(
-      "Invalid user-data-dir config.  Either must specify no " +
+      "Invalid --user-data-dir config.  Either must specify no " +
         "user-data-dir, or a user-data-dir argument for a directory that " +
         "does not currently exist, or specify a valid user-data-dir (with " +
         "the specified profile as a subdirectory in the user-data-dir if " +
@@ -382,12 +386,16 @@ export const runConfigForArgs = async (args: Namespace): Promise<RunConfig> => {
   const outputPath = args.output as undefined | "string";
   const outputHandle = await handleForResults(outputPath, args.url);
 
+  assert(typeof args.preserve_pages === "boolean");
+  const preservePages = args.preserve_pages;
+
   return {
     binary: binaryPath,
     browser: browserType,
     loggingLevel: loggingLevel,
     measurements: mesToPerform,
     output: outputHandle,
+    preservePages: preservePages,
     profile: validatedProfile,
     seconds: args.seconds,
     timeout: args.timeout,