+++ src/measure.ts
@@ -8,11 +8,13 @@ import {
   MeasurementResult,
   BaseMeasurerChild,
 } from "./measurements/base.js";
+import { MemoryMeasurer } from "./measurements/memory.js";
 import { NetworkMeasurer } from "./measurements/network.js";
 import { TimingMeasurer } from "./measurements/timing.js";
 import { MeasurementType, Report } from "./types.js";
 
 const measurerTypeToClassMap: Record<MeasurementType, BaseMeasurerChild> = {
+  [MeasurementType.Memory]: MemoryMeasurer,
   [MeasurementType.Network]: NetworkMeasurer,
   [MeasurementType.Timing]: TimingMeasurer,
 };
@@ -24,39 +26,60 @@ export const measureURL = async (
   seconds: number,
   timeout: number,
   measurements: MeasurementType[],
+  preservePages: boolean,
 ): Promise<Report> => {
+  const log = logger.prefixedLogger("Measure(): ");
+  const prevPages = context.pages();
+  const numPrevPages = prevPages.length;
+  log.verbose(`pages from previous session: ${numPrevPages.toString()} pages.`);
+
+  // Unless we're preserving existing pages, we need to force things to
+  // start in a clean profile by closing all open pages, and then
+  // re-enabling networking .
+  if (!preservePages) {
+    for (const aPage of prevPages) {
+      const pageUrl = aPage.url();
+      log.verbose("closing page: " + pageUrl);
+      await aPage.close();
+      log.verbose("page closed: " + pageUrl);
+    }
+    log.verbose("Re-enabling network for context.");
+    await context.setOffline(false);
+    log.verbose("Networking re-enabled.");
+  }
+
   // Create and instantiate any measurement classes that were requested
   const measurers = new Map<MeasurementType, BaseMeasurer>();
   for (const aMeasurementType of measurements) {
     const aMeasurerType = measurerTypeToClassMap[aMeasurementType];
     const aMeasurer = new aMeasurerType(logger, url, context);
     aMeasurer.instrumentContext();
-    logger.verbose("Instrumenting context for measurement: ", aMeasurementType);
+    log.verbose("Instrumenting context for measurement: ", aMeasurementType);
     measurers.set(aMeasurementType, aMeasurer);
   }
 
-  logger.verbose("Creating empty page (i.e., new tab).");
+  log.verbose("Creating empty page (i.e., new tab).");
   const page = await context.newPage();
 
   const startTime = new Date();
-  logger.info(`Navigating to url="${page.url()}"`);
+  log.info(`Navigating to url="${page.url()}"`);
   const navRequest = await page.goto(url.toString(), {
     timeout: timeout * 1000,
     waitUntil: "commit",
   });
   assert(navRequest);
 
-  logger.info(`Arrived at url="${page.url()}"`);
-  logger.info(`Letting page load for "${String(seconds)}" seconds`);
+  log.info(`Arrived at url="${page.url()}"`);
+  log.info(`Letting page load for "${String(seconds)}" seconds`);
   await page.waitForTimeout(seconds * 1000);
 
   for (const aMeasurer of measurers.values()) {
-    logger.verbose("Closing measurements for: ", aMeasurer.measurementType());
+    log.verbose("Closing measurements for: ", aMeasurer.type);
     aMeasurer.close();
   }
 
   const eventDrainTimeMs: number = 5 * 1000;
-  logger.verbose(
+  log.verbose(
     `Waiting "${String(eventDrainTimeMs)}ms" for events 'in-the-air' to ` +
       "complete. (Note, they are not include in measurement amounts)",
   );
@@ -64,7 +87,7 @@ export const measureURL = async (
 
   const results = {} as Record<MeasurementType, MeasurementResult | null>;
   for (const [aMeasurementType, aMeasurer] of measurers.entries()) {
-    logger.verbose(`Collecting results for measurement "${aMeasurementType}"`);
+    log.verbose(`Collecting results for measurement "${aMeasurementType}"`);
     results[aMeasurementType] = await aMeasurer.collect();
   }